Intermediate (running)

+++ pressure.ino
@@ -18,7 +18,7 @@ template <rpacPin_t p> void rpac::Pressure <p>::setup (loggerCBs_t & lcbs) {
   //
   pinMode (static_cast <uint8_t> (p), INPUT) ;
   //
-  lcbs.add ([](void) -> unsigned long { return analogRead (static_cast <uint8_t> (p)) ; }, "Pressure PIN" + String (static_cast <int> (p))) ;
+  lcbs.add ([](void) -> unsigned long { return static_cast <uint16_t> (analogRead (static_cast <uint8_t> (p))) ; }, "Pressure PIN" + String (static_cast <uint8_t> (p))) ;
   //
 }
 //
